MatFree: refactoring into submodule
Move matrix contexts into operators submodule.
